Move iterator/ to range/
