+++ include/boost/simd/range/detail/aligned_output_iterator.hpp
@@ -6,19 +6,18 @@
   (See accompanying file LICENSE.md or copy at http://boost.org/LICENSE_1_0.txt)
 **/
 //==================================================================================================
-#ifndef BOOST_SIMD_ITERATOR_DETAIL_ALIGNED_INPUT_ITERATOR_HPP_INCLUDED
-#define BOOST_SIMD_ITERATOR_DETAIL_ALIGNED_INPUT_ITERATOR_HPP_INCLUDED
+#ifndef BOOST_SIMD_RANGE_DETAIL_ALIGNED_OUTPUT_ITERATOR_HPP_INCLUDED
+#define BOOST_SIMD_RANGE_DETAIL_ALIGNED_OUTPUT_ITERATOR_HPP_INCLUDED
 
-#include <boost/simd/function/aligned_load.hpp>
+#include <boost/simd/function/aligned_store.hpp>
 #include <boost/simd/pack.hpp>
 #include <boost/simd/meta/cardinal_of.hpp>
-#include <boost/simd/iterator/detail/uncheck_iterator.hpp>
-#include <boost/simd/iterator/detail/input_iterator_base.hpp>
+#include <boost/simd/range/detail/uncheck_iterator.hpp>
+#include <boost/simd/range/detail/output_iterator_base.hpp>
 #include <boost/iterator/iterator_adaptor.hpp>
-#include <boost/align/is_aligned.hpp>
 #include <boost/core/ignore_unused.hpp>
-#include <boost/config.hpp>
-#include <iterator>
+#include <boost/align/is_aligned.hpp>
+#include <boost/assert.hpp>
 
 namespace boost { namespace simd { namespace detail
 {
@@ -29,25 +28,25 @@ namespace boost { namespace simd { namespace detail
                                   >
                             >::value
           >
-  struct  aligned_input_iterator
-        : details::input_iterator_base
+  struct  aligned_output_iterator
+        : details::output_iterator_base
                   < Iterator
                   , C
                   , typename std::iterator_traits<Iterator>::value_type
-                  , tag::aligned_load_
+                  , tag::aligned_store_
                   >
   {
-    typedef details::input_iterator_base
+    typedef details::output_iterator_base
                   < Iterator
                   , C
                   , typename std::iterator_traits<Iterator>::value_type
-                  , tag::aligned_load_
-                  >                                           parent;
+                  , tag::aligned_store_
+                  >                                               parent;
 
-    BOOST_FORCEINLINE aligned_input_iterator() : parent() {}
-    BOOST_FORCEINLINE aligned_input_iterator( parent const& src ) : parent(src) {}
+    BOOST_FORCEINLINE aligned_output_iterator() : parent() {}
+    BOOST_FORCEINLINE aligned_output_iterator( parent const& src ) : parent(src) {}
 
-    explicit BOOST_FORCEINLINE aligned_input_iterator(Iterator p) : parent(p)
+    explicit BOOST_FORCEINLINE aligned_output_iterator(Iterator p) : parent(p)
     {
       // MSVC SCL_SECURE mode adds extra-check that make aligned end
       // difficult to check with this.
@@ -57,34 +56,34 @@ namespace boost { namespace simd { namespace detail
       using target_pack = pack<typename std::iterator_traits<Iterator>::value_type, C>;
       BOOST_ASSERT_MSG
       ( boost::alignment::is_aligned(&(*lp) , target_pack::alignment )
-      , "The constructor of aligned_input_iterator has been called on a pointer "
+      , "The constructor of iterator<T,C> has been called on a pointer "
         "which alignment is not compatible with the current SIMD extension."
       );
     }
   };
 
   template<typename Iterator>
-  BOOST_FORCEINLINE aligned_input_iterator<Iterator> aligned_input_begin(Iterator p)
+  BOOST_FORCEINLINE aligned_output_iterator<Iterator> aligned_output_begin(Iterator p)
   {
-    return aligned_input_iterator<Iterator>(p);
+    return aligned_output_iterator<Iterator>(p);
   }
 
   template<std::size_t C, typename Iterator>
-  BOOST_FORCEINLINE aligned_input_iterator<Iterator, C> aligned_input_begin(Iterator p)
+  BOOST_FORCEINLINE aligned_output_iterator<Iterator,C> aligned_output_begin(Iterator p)
   {
-    return aligned_input_iterator<Iterator, C>(p);
+    return aligned_output_iterator<Iterator,C>(p);
   }
 
   template<typename Iterator>
-  BOOST_FORCEINLINE aligned_input_iterator<Iterator> aligned_input_end(Iterator p)
+  BOOST_FORCEINLINE aligned_output_iterator<Iterator> aligned_output_end(Iterator p)
   {
-    return aligned_input_iterator<Iterator>(p);
+    return aligned_output_iterator<Iterator>(p);
   }
 
   template<std::size_t C, typename Iterator>
-  BOOST_FORCEINLINE aligned_input_iterator<Iterator,C> aligned_input_end(Iterator p)
+  BOOST_FORCEINLINE aligned_output_iterator<Iterator,C> aligned_output_end(Iterator p)
   {
-    return aligned_input_iterator<Iterator, C>(p);
+    return aligned_output_iterator<Iterator,C>(p);
   }
 } } }
