+++ include/boost/simd/range/detail/input_iterator_base.hpp
@@ -6,8 +6,8 @@
   (See accompanying file LICENSE.md or copy at http://boost.org/LICENSE_1_0.txt)
 **/
 //==================================================================================================
-#ifndef BOOST_SIMD_ITERATOR_DETAILS_INPUT_ITERATOR_BASE_HPP_INCLUDED
-#define BOOST_SIMD_ITERATOR_DETAILS_INPUT_ITERATOR_BASE_HPP_INCLUDED
+#ifndef BOOST_SIMD_RANGE_DETAILS_INPUT_ITERATOR_BASE_HPP_INCLUDED
+#define BOOST_SIMD_RANGE_DETAILS_INPUT_ITERATOR_BASE_HPP_INCLUDED
 
 #include <boost/simd/pack.hpp>
 #include <boost/simd/meta/cardinal_of.hpp>
