+++ include/boost/simd/range/detail/output_iterator.hpp
@@ -6,13 +6,13 @@
   (See accompanying file LICENSE.md or copy at http://boost.org/LICENSE_1_0.txt)
 **/
 //==================================================================================================
-#ifndef BOOST_SIMD_ITERATOR_DETAIL_OUTPUT_ITERATOR_HPP_INCLUDED
-#define BOOST_SIMD_ITERATOR_DETAIL_OUTPUT_ITERATOR_HPP_INCLUDED
+#ifndef BOOST_SIMD_RANGE_DETAIL_OUTPUT_ITERATOR_HPP_INCLUDED
+#define BOOST_SIMD_RANGE_DETAIL_OUTPUT_ITERATOR_HPP_INCLUDED
 
 #include <boost/simd/function/store.hpp>
 #include <boost/simd/pack.hpp>
 #include <boost/simd/meta/cardinal_of.hpp>
-#include <boost/simd/iterator/detail/output_iterator_base.hpp>
+#include <boost/simd/range/detail/output_iterator_base.hpp>
 #include <boost/iterator/iterator_adaptor.hpp>
 #include <iterator>
