+++ test/api/CMakeLists.txt
@@ -9,5 +9,5 @@
 add_subdirectory(helper)
 add_subdirectory(meta)
 add_subdirectory(pack)
-add_subdirectory(iterator)
+add_subdirectory(range)
 add_subdirectory(algorithm)
